Move IndexArms action to impl module
